Rename ItemCreated, ItemDeleted, and move ItemChanged.
Change CreatedItem signal to ItemCreated and DeletedItem to ItemDeleted,
as suggested by Kevin Krammer.

Also move 'changed' signal from Item interface to ItemChanged on
Collection interface. If people want to monitor properties of an
individual item they can do so with the standard dbus properties
api. ItemChanged is useful for monitoring changes in a whole
collection (in a GUI for example).