[daemon] Rework signal handling, startup procedure.
Reworking the signal handling, so we have a specific thread for
handling signals. All other threads ignore signals. This is the
recommended way to 'mix' signals and threads.

In addition we move the part of startup that we can perform
after forking into separate procedures. In particular we try
to prevent starting threads until after we've forked. Besides
the obvious sanity, this helps the signal handling to work
for all threads.