Use default D-Bus timeout when finding daemon.
When we're trying to find the keyring daemon over D-Bus it's likely
that we're service activating it. If we've got a timeout of one
second and the system is under heavy load (booting, for example), then
it will not start in time, and gnome_keyring_is_available() will
return FALSE.

The default timeout in dbus-daemon is 25 seconds at the time of
writing. Although this sounds a lot for a blocking call in
_is_available, the only situation where it could take this amount of
time is if the keyring daemon is on the bus but is not responding
(it's wedged itself somehow). In all other circumstances, a response
should come immediately.

Signed-off-by: Jonny Lamb <jonnylamb@gnome.org>
Signed-off-by: Stef Walter <stef@memberwebs.com>