[daemon] Fix a delay when the daemon quits
g-k-d is a multi-threaded program, using a signal thread to stop the
main loop. The main thread was using raise() to send SIGTERM to the
signal thread, but it actually sends the signal to itself;
pthread_kill() has to be used.

This also implies a switch to pthread_create() instead of
g_thread_create() since the GThread API doesn't cover pthread_kill().