[login] Fix security problem storing user's login password in keyring.
Don't store the user's login password in the keyring, which was being
done when initializing the PKCS#11 user storage module.

In addition put in a temporary fix which removes such passwords stored.
This problem started during the 2.29.x betas, so anyone doing beta testing
will be covered by this removal.

Fixes bug #616174