[dbus] Send a content-type along with secret as per spec.
Recent changes to the secret service spec say we should be sending
a content-type along with the secret. This changes the signature of
a secret structure to (oayays).

We have nowhere to store this content-type so we simply accept and
ignore it, and return a place holder value where necessary.

This is an incompatible change to the dbus specification and both
gnome-keyring and libgnome-keyring need to be upgraded in sync.