Set correct daemon SELinux context when started from pam module
With confined users we want to run the gnome-keyring-daemon as
sgkeyringd_staff_t (Fedora case), but currently this does not
work properly from the pam module.

This patch will fix the problem.  We are ignoring error conditions
since for most users the errors will not effect unconfined users,
and on a confined user it would still partially work but generate
an AVC.

Patch by Daniel Walsh <dwalsh@redhat.com>
https://bugzilla.redhat.com/show_bug.cgi?id=684225