buffer_get_time: use time_t as last argument
ctime and mtime are time_t values.
However, buffer_get_time() takes glong as its last argument.
This is problematic because depending on the operating system and arch,
time_t can be different.
For e.g. on OpenBSD it is both long long for 32-bit and 64-bit
architectures while on Linux, it is long for 32-bit.

https://bugzilla.gnome.org/show_bug.cgi?id=706405