gkm-gnome2-file: Fix leaks in create_cipher()
'key' and 'iv' were allocated before calling
egg_symkey_generate_simple() but this function allocates the memory
needed for the 'key' and 'iv' return value, so the memory which was
allocated in create_cipher() is lost and leaked.
This also uses egg_secure_memory_free() to free 'key' memory as
egg_symkey_generate_simple() allocates it with egg_secure_alloc().

https://bugzilla.gnome.org/show_bug.cgi?id=738508