xdg: Don't leak ref in lookup_or_create_assertion_key()
When lookup_or_create_assertion_key() creates a new assertion key,
it own a reference on the GBytes created by create_assertion_key()
and it will then take an extra ref on it when associating it with a
GkmAssertion instance.

This will cause a leak of the memory returned by create_assertion_key()
as one of these references will never be dropped.

https://bugzilla.gnome.org/show_bug.cgi?id=738508