Fix PKCS#11 RPC handling of multi-threaded clients
In a multi-threaded client which makes concurrent calls into the PKCS#11
module (which is permitted as long as it's for separate sessions), we
might end up with multiple RPC connections to the dæmon.

These are currently being treated as independent client applications,
which leads to spurious CKR_USER_NOT_LOGGED_IN and other errors as
described at https://bugzilla.redhat.com/show_bug.cgi?id=1173279

Fix this by moving the CK_G_APPLICATION state out of the CallState and
into a separate refcounted ClientInstance data structure, and looking up
the appropriate ClientInstance for new connections before creating a new
one.

The ClientInstance is accessed without locking during certain calls
because we know it *cannot* go away while a refcount is still held on it.