ssh-agent: Propagate stderr to journal when spawning ssh-add
ssh-add fails in certain occasions, such as when the file permissions
of private key is not unsafe.  To help diagnostics, propagate the
stderr output from the command to journal.

As the ssh commands send error message with trailing CR for each line,
we need to scrub it so as not to confuse journald.

https://bugzilla.gnome.org/show_bug.cgi?id=794361