ssh-agent: Use the same parameters for accessing login keyring
When looking up a secret in the login keyring, do not supply any
schema in the criteria, while using "org.freedesktop.Secret.Generic"
as schema when storing it.  This is for backward compatibility with
gnome-keyring 2.29, which used "org.gnome.keyring.EncryptionKey" as
schema.

In addtion, use the same label for the newly stored passwords as
before.

https://bugzilla.gnome.org/show_bug.cgi?id=794368