Remove tap-gtester, rely on GLib 2.38+ built-in TAP output instead
GLib 2.62+ outputs TAP by default, which breaks the tap-gtester
script's expectations. In older GLib versions since 2.38, --tap is an
opt-in to the TAP output mode, so use that.

This is essentially the same issue as
https://github.com/cockpit-project/cockpit/pull/11998 (cockpit was the
source of the tap-gtester script), GNOME/gcr!19, and GNOME/libsecret!5.

Signed-off-by: Simon McVittie <smcv@debian.org>
Bug-Debian: https://bugs.debian.org/940157