Default "automatically unlock" checkbox to false
The GCR prompt used to unlock keyrings and other secure objects has a
checkbox for automatically unlocking the object when the user logs in.
Previously, this checkbox was checked by default if the object in
question had the CKA_GNOME_TRANSIENT attribute set. This linkage frankly
makes no sense; the change can be traced back to a nine-year-old commit
(084ce5e08) which appears to have mistakenly altered code that used
CKA_GNOME_TRANSIENT to determine the *nature* of the choice into code
that used CKA_GNOME_TRANSIENT to set the default *value* of the choice.

Defaulting this checkbox to true, which was the practical effect of
using CKA_GNOME_TRANSIENT in this way, has obvious drawbacks for
security, as the reporter of #7 and the various reporters in bugs linked
from there have attested.

Removing this linkage is a small change, but it results in a larger
amount of code being no longer needed, which this commit cleans up.

Fixes #7.