egg-secure-memory: Allocate enough space for the guard pointers
Cell allocations have an overhead of two pointer-sized words, which we
need to take into account. Otherwise, whenever we try to allocate a new
block to hold at least `DEFAULT_BLOCK_SIZE - 2 * sizeof (void *)` bytes,
it will end up too small to allocate a cell of that size and we will
fall back to non-mlock'd memory with a warning.

Fixes: 982aadd9 "Completely overhaul the secure memory allocator"
Resolves: https://gitlab.gnome.org/GNOME/gnome-keyring/-/issues/36
Signed-off-by: Simon McVittie <smcv@debian.org>