doap: Add myself as maintainer, remove Stef
Stef has been inactive on GNOME development for quite a while now, and
the email address listed in the doap is no longer in use even, so remove
him.

That being said, I'm probably the closest to a co-maintainer with Daiki
as I'm maintaining libsecret anyway, so I get to deal with gnome-keyring
whether I like it or not.