ssh-agent: stop waiting for agent to become ready if it's dead
If the ssh-agent process we launch dies before it becomes ready to take
input, self->pid will be set to 0 by on_child_watch().  If that happens,
there's no point in continuing to wait for the process to become ready.
This should avoid an unnecessary five-second wait in cases like [1] or
[2].

[1] https://bugzilla.gnome.org/show_bug.cgi?id=794848
[2] https://bugzilla.redhat.com/show_bug.cgi?id=1841855 (migrated to
https://issues.redhat.com/browse/RHEL-9302)