pkcs11: Fix memory leaks in test-sexp
This commit fixes the memory leaking of the "base-sexp" of the generated
keys in our tests. The main problem in trying to solve this is that
`GkmSexp` takes ownership of the underlying `gcry_sexp_t`, so if we just
unref the `GkmSexp` at the end of the function, we would end up with a
double free.

We fix this here by explicitly NULL-ing the pointer, either by using
`g_steal_pointer()`, or by just setting it to NULL at the end.

To make things a bit easier also, this commit adds `g_autoptr`
functionality to `GkmSexp`.