dbus: Set xdg:schema for Secret Portal secrets
There's currently no good way of distinguishing application master
secrets created for the XDG Secret Portal backend.

We've added a user-friendly translatable label in commit cd3128488f, but
that's still not something one can use to deterministically detect this
case. So, add a `xdg:schema` attribute to the secret's fields.